fuz.merge: This package provides functions to implement and aid in Fuzzy Matching

This package lets you implement fuzzy merging so you can merge datasets when the values to be matched aren't exactly the same. It also includes a clean function to remove prefixes, sufixes, and other sequences of characters which the user wants to remove. It also includes a function, scramble(), which will change some of the characters of the selected column to allow for testing of the package.

See also